Rename MathUtil tests to ArrayUtil tests
